apps/services/alert-service/requirements.txt
# SAHOOL Alert Service Dependencies
# متطلبات خدمة التنبيهات
# Updated: December 2025
# Version: 16.0.0

# Web Framework (matching shared versions)
fastapi==0.126.0
uvicorn[standard]==0.34.0
pydantic==2.10.3

# HTTP Client
httpx==0.28.1

# Database & ORM
tortoise-orm==0.21.7
asyncpg==0.30.0

# NATS Messaging
nats-py==2.9.0

# Redis for caching and pub/sub
redis==5.2.1

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2

# APScheduler for alert expiration checks
apscheduler==3.10.4

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==4.1.0
